feat: Admin全featureのGraphQLクエリにUser name/emailフィールドを追加
スキーマ・コード生成を更新し、classes/competitions/images/information/
locations/sports/tags/usersの各featureでユーザー名・メール表示に対応
