+++ apps/admin/src/features/classes/hooks/useClasses.ts
@@ -1,8 +1,11 @@
+import { useMemo } from 'react'
 import { useGetAdminGroupsForClassesQuery } from '@/gql/__generated__/graphql'
+import { useFilterParams } from '@/hooks/useFilterParams'
 import type { Class } from '../types'
 
 export function useClasses() {
   const { data, loading, error } = useGetAdminGroupsForClassesQuery()
+  const { values: fp, set: setFilter, reset: resetFilters } = useFilterParams(['keyword'])
 
   const classes: Class[] = (data?.groups ?? []).map(g => ({
     id: g.id,
@@ -10,5 +13,13 @@ export function useClasses() {
     memberCount: g.users.length,
   }))
 
-  return { data: classes, loading, error: error ?? null }
+  const keyword = fp.keyword
+
+  const filtered = useMemo(() => {
+    if (!keyword) return classes
+    const kw = keyword.toLowerCase()
+    return classes.filter(c => c.name.toLowerCase().includes(kw))
+  }, [classes, keyword])
+
+  return { data: filtered, allData: classes, keyword, setFilter, resetFilters, loading, error: error ?? null }
 }
