+++ apps/admin/src/features/competitions/components/CompetitionListPage.tsx
@@ -1,4 +1,4 @@
-import { useMemo, useCallback } from 'react'
+import { useMemo } from 'react'
 import { Box, Button, Card, CardContent, CircularProgress, Typography } from '@mui/material'
 import AddIcon from '@mui/icons-material/Add'
 import { useCompetitions } from '../hooks/useCompetitions'
@@ -6,7 +6,6 @@ import { QueryError } from '@/components/ui/QueryError'
 import { CompetitionCard } from './CompetitionCard'
 import { CARD_GRADIENT, ACTION_BUTTON_SX } from '@/styles/commonSx'
 import { SearchFilterBar, type FilterDef } from '@/components/ui/SearchFilterBar'
-import { useFilterParams } from '@/hooks/useFilterParams'
 import { DragHandle } from '@/components/ui/DragHandle'
 import { useDisplayOrderDnD } from '@/hooks/useDisplayOrderDnD'
 import { useUpdateAdminCompetitionsDisplayOrderMutation } from '@/gql/__generated__/graphql'
@@ -22,12 +21,22 @@ const TYPE_OPTIONS = [
 ]
 
 export function CompetitionListPage({ onNavigateToCreate, onSelectCompetition }: Props) {
-  const { data: competitions, loading, error } = useCompetitions()
-  const { values: fp, set: setFilter, reset: resetFilters } = useFilterParams(['keyword', 'type', 'sport', 'scene'])
-  const keyword = fp.keyword
-  const typeFilter = fp.type
-  const sportFilter = fp.sport
-  const sceneFilter = fp.scene
+  const {
+    data: filtered,
+    allData: competitions,
+    keyword,
+    typeFilter,
+    sportFilter,
+    sceneFilter,
+    sportOptions,
+    sceneOptions,
+    setFilter,
+    handleFilterChange,
+    resetFilters,
+    loading,
+    error,
+  } = useCompetitions()
+
   const hasFilter = !!(keyword || typeFilter || sportFilter || sceneFilter)
 
   const [reorderMutation] = useUpdateAdminCompetitionsDisplayOrderMutation({
@@ -44,22 +53,6 @@ export function CompetitionListPage({ onNavigateToCreate, onSelectCompetition }:
     handleDragEnd,
   } = useDisplayOrderDnD(competitions, (input) => reorderMutation({ variables: { input } }))
 
-  const sportOptions = useMemo(() => {
-    const map = new Map<string, string>()
-    for (const c of competitions) {
-      if (!map.has(c.sportId)) map.set(c.sportId, c.sportName)
-    }
-    return Array.from(map, ([value, label]) => ({ value, label })).sort((a, b) => a.label.localeCompare(b.label))
-  }, [competitions])
-
-  const sceneOptions = useMemo(() => {
-    const set = new Set<string>()
-    for (const c of competitions) {
-      if (c.sceneName) set.add(c.sceneName)
-    }
-    return Array.from(set).sort().map(name => ({ value: name, label: name }))
-  }, [competitions])
-
   const filterDefs: FilterDef[] = useMemo(() => [
     { key: 'type', label: 'タイプ', options: TYPE_OPTIONS },
     { key: 'sport', label: '競技', options: sportOptions },
@@ -71,26 +64,6 @@ export function CompetitionListPage({ onNavigateToCreate, onSelectCompetition }:
     sport: sportFilter,
     scene: sceneFilter,
   }), [typeFilter, sportFilter, sceneFilter])
-
-  const handleFilterChange = useCallback((key: string, value: string) => {
-    setFilter(key, value)
-  }, [setFilter])
-
-  const filtered = useMemo(() => {
-    let result = competitions
-    if (keyword) {
-      const kw = keyword.toLowerCase()
-      result = result.filter(c =>
-        c.name.toLowerCase().includes(kw) ||
-        c.sportName.toLowerCase().includes(kw) ||
-        c.sceneName.toLowerCase().includes(kw)
-      )
-    }
-    if (typeFilter) result = result.filter(c => c.type === typeFilter)
-    if (sportFilter) result = result.filter(c => c.sportId === sportFilter)
-    if (sceneFilter) result = result.filter(c => c.sceneName === sceneFilter)
-    return result
-  }, [competitions, keyword, typeFilter, sportFilter, sceneFilter])
 
   const renderItems = hasFilter ? filtered : displayItems
