+++ apps/admin/src/features/locations/components/LocationCreatePage.tsx
@@ -22,9 +22,9 @@ export function LocationCreatePage({ onBack, onSave }: Props) {
   }
 
   return (
-    <Box>
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 3 }}>
       <BackButton onClick={onBack} />
-      <Breadcrumbs separator="/" sx={{ mb: 2 }}>
+      <Breadcrumbs separator="/" sx={{ mb: 0 }}>
         <ButtonBase onClick={onBack} sx={BREADCRUMB_LINK_SX}>
           場所
         </ButtonBase>
