+++ apps/admin/src/features/locations/components/LocationDetailPage.tsx
@@ -50,9 +50,9 @@ export function LocationDetailPage({ locationId, onBack, onSave, onDelete }: Pro
   }
 
   return (
-    <Box>
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 3 }}>
       <BackButton onClick={onBack} />
-      <Breadcrumbs separator="/" sx={{ mb: 2 }}>
+      <Breadcrumbs separator="/" sx={{ mb: 0 }}>
         <ButtonBase onClick={onBack} sx={BREADCRUMB_LINK_SX}>
           場所
         </ButtonBase>
