+++ apps/admin/src/features/tags/components/TagDetailPage.tsx
@@ -38,9 +38,9 @@ export function TagDetailPage({ tagId, onBack }: Props) {
   }
 
   return (
-    <Box>
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 3 }}>
       <BackButton onClick={onBack} />
-      <Breadcrumbs separator="/" sx={{ mb: 2 }}>
+      <Breadcrumbs separator="/" sx={{ mb: 0 }}>
         <ButtonBase onClick={onBack} sx={BREADCRUMB_LINK_SX}>
           タグ
         </ButtonBase>
