+++ apps/admin/src/features/users/types.ts
@@ -2,6 +2,7 @@ export type User = {
   id: string
   name: string
   email: string
+  studentNumber: string
   role: string
   groupName: string
   teams: { id: string; name: string }[]
